Add CLI
Mostly mimicking what `happo compare` does in the regular happo.io CLI
code. For now the api key and secret are passed in. I might switch to
reading the .happo.js file at some point instead.

Or I might scrap this library completely and bake it in with the
happo.io CLI.